Move geocoding stuff to dedicated source file.
